// src/app/model/coin.ts
export class Coin {

    Id: string;
    Url: string;
    ImageUrl: string;
    Name: string;
    Symbol: string;
    CoinName: string;
    FullName: string;
    Algorithm: string;
    ProofType: string;
    FullyPremined: string;
    TotalCoinSupply: string;
    PreMinedValue: string;
    TotalCoinsFreeFloat: string;
    SortOrder: string;
    Sponsored: boolean;

    constructor() {

    }

}